requirements.txt
pygame-ce>=2.5
numpy>=1.24
scipy>=1.11
